feat: move NMDA to abstract models
